FIX: pretty text allow list
Reword whitelist to allowlist in pretty-text.
This library is used by plugins so we need deprecation notice.
